requirements.txt
asgiref==3.6.0
Django==3.2
django-filter==22.1
django-tinymce==3.5.0
djangorestframework==3.14.0
Markdown==3.4.1
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.3
